+++ src/tasks/task2/fragments/sidebar.py
@@ -16,8 +16,8 @@ def show_sidebar():
             a = int(st.number_input('A:', value=0))
             StateVar.LEFT_BOUNDARY.set(a)
 
-            number_of_nodes = int(st.number_input('Количество узлов:', value=16, min_value=2))
-            StateVar.NUMBER_OF_NODES.set(number_of_nodes)
+            number_of_points = int(st.number_input('Количество узлов:', value=16, min_value=2))
+            StateVar.NUMBER_OF_POINTS.set(number_of_points)
 
         with right_column:
             b = int(st.number_input('B:', value=1))
@@ -26,9 +26,9 @@ def show_sidebar():
             polynomial_degree = int(
                 st.number_input(
                     'Степень многочлена:',
-                    value=min(7, number_of_nodes - 1),
+                    value=min(7, number_of_points - 1),
                     min_value=0,
-                    max_value=number_of_nodes - 1,
+                    max_value=number_of_points - 1,
                 ),
             )
             StateVar.POLYNOMIAL_DEGREE.set(polynomial_degree)
